requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
Django==3.2.5
django-ckeditor==6.1.0
django-js-asset==1.2.2
html5lib==1.1
mechanize==0.4.5
psycopg2-binary==2.9.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
webencodings==0.5.1
